requirements.txt
fastapi==0.98.0
uvicorn==0.22.0
neo4j==5.9.0
python-dotenv==1.0.0